server/requirements.txt
websockets~=12.0
aiohttp[speedups]~=3.9
argon2-cffi~=23.1
bidict~=0.23
